requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.23
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
smmap==5.0.2
typing_extensions==4.15.0
urllib3==2.6.3
black